refactored TitlesSpec into ReportRewriteRuleSpec
